fix: pengajuan only for one investor

+++ src/migrations/20260417153530_investasi.js
@@ -20,7 +20,7 @@ exports.up = function (knex) {
       .references("negosiasis.id")
       .onDelete("SET NULL");
 
-    table.bigint("nominal_investasi", 20, 2).unsigned().notNullable();
+    table.decimal("nominal_investasi", 20, 2).unsigned().notNullable();
     table.decimal("return_investasi", 20, 2).unsigned().notNullable();
     table.timestamps(true, true);
   });
